module: audio: circ_buf: return sample count as size_t
Change the return type of cir_buf_samples_without_wrap_s16() and
cir_buf_samples_without_wrap_s32() from int to size_t, and update the
sample-count variables in all callers (volume, mixer, aria, asrc).

The returned value is a non-negative element count derived from pointer
arithmetic, so size_t matches the rest of the circular-buffer API.

+++ src/audio/volume/volume_hifi5.c
@@ -65,7 +65,7 @@ static void vol_s24_to_s24_s32(struct processing_module *mod, struct cir_buf_sou
 	ae_int32x2 volume, volume1;
 	ae_int32x4 *buf;
 	ae_int32x4 *buf_end;
-	int i, n, m;
+	size_t i, n, m;
 	ae_int32x4 *vol;
 	ae_valignx2 inu;
 	ae_valignx2 outu = AE_ZALIGN128();
@@ -73,7 +73,7 @@ static void vol_s24_to_s24_s32(struct processing_module *mod, struct cir_buf_sou
 	ae_int32x4 *out = sink->ptr;
 	const int channels_count = cd->channels;
 	const int inc = sizeof(ae_int32x4);
-	int samples = channels_count * frames;
+	size_t samples = channels_count * frames;
 
 	/* to ensure the address is 16-byte aligned and avoid risk of
 	 * error loading of volume gain while the cd->vol would be set
@@ -149,12 +149,12 @@ static void vol_passthrough_s24_to_s24_s32(struct processing_module *mod,
 {
 	struct vol_data *cd = module_get_private_data(mod);
 	ae_int32x2 in_sample, in_sample1;
-	int i, n, m;
+	size_t i, n, m;
 	ae_valignx2 inu;
 	ae_valignx2  outu = AE_ZALIGN128();
 	const ae_int32x4 *in = source->ptr;
 	ae_int32x4 *out = sink->ptr;
-	int samples = cd->channels * frames;
+	size_t samples = cd->channels * frames;
 
 	while (samples) {
 		m = cir_buf_samples_without_wrap_s32(in, source->buf_end);
@@ -195,7 +195,7 @@ static void vol_s32_to_s24_s32(struct processing_module *mod, struct cir_buf_sou
 	ae_int32x2 in_sample, in_sample1;
 	ae_int32x2 out_sample, out_sample1;
 	ae_int32x2 volume, volume1;
-	int i, n, m;
+	size_t i, n, m;
 	ae_int32x4 *buf;
 	ae_int32x4 *buf_end;
 	ae_int32x4 *vol;
@@ -203,7 +203,7 @@ static void vol_s32_to_s24_s32(struct processing_module *mod, struct cir_buf_sou
 	ae_valignx2 outu = AE_ZALIGN128();
 	const int channels_count = cd->channels;
 	const int inc = sizeof(ae_int32x4);
-	int samples = channels_count * frames;
+	size_t samples = channels_count * frames;
 	const ae_int32x4 *in = source->ptr;
 	ae_int32x4 *out = sink->ptr;
 
@@ -279,11 +279,11 @@ static void vol_passthrough_s32_to_s24_s32(struct processing_module *mod,
 {
 	struct vol_data *cd = module_get_private_data(mod);
 	ae_int32x2 in_sample, in_sample1;
-	int i, n, m;
+	size_t i, n, m;
 	ae_valignx2 inu;
 	ae_valignx2 outu = AE_ZALIGN128();
 	const int channels_count = cd->channels;
-	int samples = channels_count * frames;
+	size_t samples = channels_count * frames;
 	const ae_int32x4 *in = source->ptr;
 	ae_int32x4 *out = sink->ptr;
 
@@ -325,7 +325,7 @@ static void vol_s16_to_s16(struct processing_module *mod, struct cir_buf_source 
 	ae_int32x2 out_temp, out_temp1;
 	ae_int16x4 in_sample, in_sample1;
 	ae_int16x4 out_sample, out_sample1;
-	int i, n, m;
+	size_t i, n, m;
 	ae_int32x4 *buf;
 	ae_int32x4 *buf_end;
 	ae_int32x4 *vol;
@@ -335,7 +335,7 @@ static void vol_s16_to_s16(struct processing_module *mod, struct cir_buf_source 
 	ae_int16x8 *out = sink->ptr;
 	const int channels_count = cd->channels;
 	const int inc = sizeof(ae_int32x4);
-	int samples = channels_count * frames;
+	size_t samples = channels_count * frames;
 
 	/** to ensure the address is 16-byte aligned and avoid risk of
 	 * error loading of volume gain while the cd->vol would be set
@@ -421,13 +421,13 @@ static void vol_passthrough_s16_to_s16(struct processing_module *mod,
 {
 	struct vol_data *cd = module_get_private_data(mod);
 	ae_int16x4 in_sample, in_sample1;
-	int i, n, m;
+	size_t i, n, m;
 	ae_valignx2 inu;
 	ae_valignx2 outu = AE_ZALIGN128();
 	const ae_int16x8 *in = source->ptr;
 	ae_int16x8 *out = sink->ptr;
 	const int channels_count = cd->channels;
-	int samples = channels_count * frames;
+	size_t samples = channels_count * frames;
 
 	while (samples) {
 		m = cir_buf_samples_without_wrap_s16(in, source->buf_end);
